pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ncm-cli"
version = "1.0.0"
description = "A command-line tool for searching and downloading music from Netease Cloud Music"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "NCM CLI"}
]
keywords = [
    "netease",
    "music",
    "download",
    "cli",
    "netease-cloud-music"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio",
]
dependencies = [
    "click>=8.0.0",
    "requests>=2.28.0",
    "pycryptodome>=3.18.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "types-requests>=2.28.0",
]

[project.scripts]
ncm = "ncm.cli:main"

[project.urls]
Homepage = "https://github.com/example/ncm-cli"
Repository = "https://github.com/example/ncm-cli"
Issues = "https://github.com/example/ncm-cli/issues"

[tool.setuptools.packages.find]
where = ["src"]
include = ["ncm*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=ncm --cov-report=term-missing"
